requirements.txt
fastapi
uvicorn
python-dotenv
requests
google-generativeai
pydantic
sqlalchemy
psycopg2-binary